requirements.txt
pygame-ce>=2.5.0

# Web server dependencies (for web_server.py)
flask>=3.0.0
flask-socketio>=5.3.0
simple-websocket>=1.0.0
pillow>=10.0.0
